feat: SPA contract analysis system, code review fixes, and error prevention hooks
- Add SPA reverse-analysis 3-step wizard (step1: variable extraction, step2: clause decomposition, step3: template seeding) with full backend API + frontend UI
- Fix FIRecommendModal TS2345 type error (formatBillion accepts string|number) that caused CI quality gate failure
- Add block-python-antipatterns.sh PreToolUse hook to prevent UP041/B017 ruff violations at write time
- Enhance auto-format-python.sh with Phase 3 residual error feedback to Claude
- Add pytest -o "addopts=" override in run-affected-tests.sh to avoid pytest-cov dependency errors
- Update top5-error-prevention rules with P6 reinforcement (UP041/B017) and new P9 (pytest env mismatch)
- Apply multi-round code review fixes across deal-mgmt, kiis, im backends (ruff, type hints, migrations)

+++ deal-mgmt/app/services/permit_knowledge_base.py
@@ -25,7 +25,7 @@ class PermitKBEntry:
     post_filing_days: int | None = None
     required_documents: list[str] = field(default_factory=list)
     deal_structure_filter: list[str] | None = None
-    threshold_value: float | None = None
+    threshold_value: Decimal | None = None
     notes: str | None = None
